--- Hunt the wumpus final/ASCII_Logo_Art.txt
@@ -1,22 +1,0 @@
-
-
-
-  _    _   _    _   _   _   _______                         
- | |  | | | |  | | | \ | | |__   __|                        
- | |__| | | |  | | |  \| |    | |                           
- |  __  | | |  | | | . ` |    | |                           
- | |  | | | |__| | | |\  |    | |                           
- |_|__|_|  \____/_ |_|_\_|_   |_|                           
- |__   __| | |  | | |  ____|                                
-    | |    | |__| | | |__                                   
-    | |    |  __  | |  __|                                  
-    | |    | |  | | | |____                                 
- __ |_|    |_|_ |_| |______|  __   _____    _    _    _____ 
- \ \        / / | |  | | |  \/  | |  __ \  | |  | |  / ____|
-  \ \  /\  / /  | |  | | | \  / | | |__) | | |  | | | (___  
-   \ \/  \/ /   | |  | | | |\/| | |  ___/  | |  | |  \___ \ 
-    \  /\  /    | |__| | | |  | | | |      | |__| |  ____) |
-     \/  \/      \____/  |_|  |_| |_|       \____/  |_____/ 
-                                                            
- 
-  MADE BY: JEFF, ADAM, MARK AND JOHNY [TEAM B5]
