--- Hunt the wumpus final/Game_Info.txt
@@ -1,27 +1,0 @@
-THE WUMPUS LIVES IN A CAVE WHICH HAS 20 ROOMS. THERE ARE THREE TUNNELS IN EACH ROOM LEADING TO OTHER ROOMS. (THINK OF A DODECAHEDRON)
-EACH ROOM IS DIRECTLY CONNECTED TO THREE OTHER ROOMS WHICH YOU CAN ACCESS THROUGH THE CONNECTED TUBES.
-
--HAZARDS
-BOTTOMLESS PITS: TWO ROOMS IN THE CAVE HAVE A PIT IN THEM, WHEN YOU ENTER SUCH A ROOM YOU WILL FALL, DIE AND LOSE.
-WUMPUS: THE WUMPUS IS NOT BOTHERED BY HAZARDS, HOWEVER IF YOU ENTER THE SAME ROOM AS THE WUMPUS, YOU WILL BE EATEN.
-BATS: THE BATS IN THIS CAVE WILL PICK YOU UP IF YOU ENTER THE SAME ROOM AS THEM, ONCE THAT HAPPENS THE BATS WILL DROP YOU IN A RANDOM ROOM.
-
--ABILITIES
-MOVE: YOU CAN MOVE TO A CONNECTED ROOM THROUGH A TUNNEL.
-SHOOT: YOU HAVE 5 ARRORWS WHICH YOU CAN SHOOT TO SPECIFIC ROOMS, BUT BE AWARE THAT SHOOTING AN ARROW WILL SCARE THE WUMPUS, WHICH MAKES IT MOVE.
-SENSE: WHENEVER YOU ARE TWO ROOMS AWAY FROM THE WUMPUS YOU CAN SENSE ITS PRESENCE, ALSO BATS AND PITS CAN BE SENSED ONE ROOM AWAY.
-
--EXTRA'S
-AI: ITS ALSO POSSIBLE FOR AN AI TO SOLVE THE GAME FOR YOU, WHICH TAKES BETWEEN 2 AND 4 ROUNDS TO WIN.
-
--GAME MODES
-AT THE BEGINNING YOU CAN CHOOSE BETWEEN NORMAL MODE OR CUSTOM MODE.
-RANDOM MODE IS THE DEFAULT GAME MODE IN WHICH YOU GENERATE A NEW CAVE.
-CONFIGURATION MODE MAKES IT POSSIBLE TO REPLAY THE GENERATED MAP CURRENTLY SAVED IN THE MAP.TXT FILE.
-
-IF YOU HAVE TO INTERRUPT THE COMPUTER'S RUN, THEN YOU CAN DO SO BY PRESSING CTRL + C, OR CLOSING THE TERMINAL.
-IF YOU WANT TO CHECK THE DATA THATS CURRENTLY ON THE SCREEN CLICK ON THE SCREEN TO PAUSE AND PRESS ENTER TO CONTINUE.
-
-ONCE YOU HAVE REACHED THE END STATE OF THE GAME, YOU'LL HAVE THE ABILITY TO SEE THE ROUND INFO AND ALSO THE GENERAL PLAYER STATISTICS
-
-BEST OF LUCK
