--- Hunt the wumpus final/animations.hpp
@@ -1,4 +1,0 @@
-void kill_wumpus_ani();
-void wumpus_eat_ani();
-void pitt_fall_ani();
-void empty_arrows_ani();
