+++ animations.cpp
@@ -1,0 +1,1121 @@
+#include <iostream>
+#include <chrono>
+#include <thread>
+#include <string>
+#include <windows.h>
+#include "animations.hpp"
+using namespace std;
+
+void kill_wumpus_ani(){
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 @@@@@@@@@@@@@@@@@@@@@@@@@@@(@@%               %@@%@@@@@@@@@@@@@@@@@@@@@@@%                             \n"
+"                                                                         /####,                                         \n"
+"                                                                                                                        \n"
+"                          #@&.      /@@                                                                                 \n"
+"                      #@                @@                                                                              \n"
+"                    @@        (%%%%%(     /@                                                                            \n"
+"              @ #@@    ,@@@(             @# @                                                                           \n"
+"             @@ @        @                 @@                                                                           \n"
+"                        @                   @                                                                           \n"
+"                       %                    @                                                                           \n"
+"                       #@                 #@                                                                            \n"
+"                          @             @@                                                                              \n"
+"                            (@@@((#@@*                                                                                  \n"
+"                                &@#     @@                                                                              \n"
+"                             @   @ @@   &%  @%                                                                          \n"
+"                             @   @     @@    &(                                                                         \n"
+"                              @@ @   &@@ @@@  @   /@                                                                    \n"
+"                                 @    @      ,@  @  .@                                                                  \n"
+"                                 @   @      #@                                                                          \n"
+"                                 @  (@@@@@@*                                                                            \n"
+"                               ,@  &@                                                                                   \n"
+"                              &#     ,@                                                                                 \n"
+"                             @        .@                                                                                \n"
+"                            @          @                                                                                \n"
+"                          @(           @                                                                                \n"
+"                         @            @       #@@@@@@@@@@    @@@@@@@@@@@@&@@@@                                          \n"
+"        .@@@@@     @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(1000) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 @@@@@@@@@@@@@@@@@@@@@@@@@@@(@@%               %@@%@@@@@@@@@@@@@@@@@@@@@@@%                             \n"
+"                                                                         /####,                                         \n"
+"                           .(@@@&(                                                                                      \n"
+"                     ,@@@@*        &@@@@                                                                                \n"
+"                  (@@.                  /@@,                                                                            \n"
+"              @@ @@       (@@#      .%@@@ .@@                                                                           \n"
+"            @@@@@# *@@@& @@              @@@@                                                                           \n"
+"          /@@@@         @@                @@@                                                                           \n"
+"                        @@                @@                                                                            \n"
+"                        @@*              @@                                                                             \n"
+"                         .@@           @@@                                                                              \n"
+"                            &@@@# (@@@@@@@@@,                                                                           \n"
+"                                 %@, @@     @@.                                                                         \n"
+"                        &@@@(.  .,@@ @@      /@&                                                                        \n"
+"                       @@        .@@@@@       @@                                                                        \n"
+"                       @@@@@@@&&&&@@@@@     ,@@@  @@@                                                                   \n"
+"                                 ,@( @@       @@ &@@&                                                                   \n"
+"                                  @@ &@       @@                                                                        \n"
+"                                  @@ /@*     @@*                                                                        \n"
+"                                  @@&@@@@@@@@                                                                           \n"
+"                                %@&&@@                                                                                  \n"
+"                               %@*   @@                                                                                 \n"
+"                              @@,     .@@                                                                               \n"
+"                             @@         @@                                                                              \n"
+"                            @@.         /@,                                                                             \n"
+"                           *@%          &@.   #@@@@@@@@@@    @@@@@@@@@@@@&@@@@                                          \n"
+"        .@@@@@     @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(1000) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 @@@@@@@@@@@@@@@@@@@@@@@@@@@(@@%               %@@%@@@@@@@@@@@@@@@@@@@@@@@%                             \n"
+"                                                                         /####,                                         \n"
+"                                                                                                                        \n"
+"                        #@@@@@@@@@@@&                                                                                   \n"
+"                    %@@#             ,@@@                                                                               \n"
+"                  @@&     &@@@*     /@@@&@@                                                                             \n"
+"                @@%     *@%             @@@@                                                                            \n"
+"               @@   %@@@@@               @@                                                                             \n"
+"              &@@@@/   .@%               @@                                                                             \n"
+"           @@@@@        @@               &@(                                                                            \n"
+"           @@@@@@       .@@             @@@,                                                                            \n"
+"           .@@#            @@@@&&&&@@@@@@@  @@@                                                                         \n"
+"           .*                    @@   (@#     @@/                      @@@@@%                                           \n"
+"                            &@@@@@@@   @@      #@(            .,,*////(@@@@@@@@@@&                                      \n"
+"                       ,@@@#     .@&@@ #@*      @@     ,,.             @@@@@@@@@                                        \n"
+"                     %@@         (@,  #@@@      @@                     @@@@&                                            \n"
+"                    @@           @@      @@    #@/                                                                      \n"
+"                                 @@      @@ *@@@                                                                        \n"
+"                                 @@   .*@@@@                                                                            \n"
+"                                 ,@%                                                                                    \n"
+"                                 @@@@                                                                                   \n"
+"                                @@  @@*                                                                                 \n"
+"                               @@    .@@                                                                                \n"
+"                              @@       @@                                                                               \n"
+"                             @@         @@(                                                                             \n"
+"                            @@           *@@  #@@@@@@@@@@    @@@@@@@@@@@@&@@@@                                          \n"
+"        .@@@@@     @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 @@@@@@@@@@@@@@@@@@@@@@@@@@@(@@%               %@@%@@@@@@@@@@@@@@@@@@@@@@@%                             \n"
+"                                                                         /####,                                         \n"
+"                                                                                                                        \n"
+"                       &@@@@@@@@@@@@@*                                                                                  \n"
+"                   &@@#      (&@@@#   @@@,                                                                              \n"
+"                 #@@      @@(       %@@@,@@                                                                             \n"
+"                @@      *@&             @@@&                                                                            \n"
+"               @@     /@@@               @@                                                                             \n"
+"              ,@%#@@@@  @@               *@/                                                                            \n"
+"           @@@@@@       %@/              @@                                                                             \n"
+"           .@@@@@@       ,@@            @@@@@%                                                                          \n"
+"           %@&@@            @@@@@@@@@@@@@     @@                                                         @@@            \n"
+"           @*                    (@,    @@     @@.                                                       @@@@@@@@.      \n"
+"                   &.            (@,    %@@     @@                                    @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@   \n"
+"                    *@@@@&&&&@@@@@@@@@@&  @@   &@.                                                       @@@@@@@@.      \n"
+"                                 @@        @@ @@.                                                        @@@@           \n"
+"                                 @@         @@@                                                                         \n"
+"                                 @@   .@@@@@/                                                                           \n"
+"                                 (@&                                                                                    \n"
+"                                 @@@@                                                                                   \n"
+"                               *@&  #@%                                                                                 \n"
+"                               @@     @@                                                                                \n"
+"                              @@       @@                                                                               \n"
+"                             /@(        @@                                                                              \n"
+"                             @@          @@                                                                             \n"
+"                            @@           @@   #@@@@@@@@@@    @@@@@@@@@@@@&@@@@                                          \n"
+"        .@@@@@     @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"       .*@@@@@@@@@@@@&(,                 /@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(          *%@@@@@@@@@@%  \n"
+"                                                                      .@@@@@@@@@@@@@@%                                  \n"
+"                                                                                      (@@@@@@@@@@@@@@@@@%,              \n"
+"                                                                                 ,@@@@,                  *@@@@@.        \n"
+"                                                                                @@&      @@@@#                ,@@@      \n"
+"                                                                               ,@@       @@@@@,                 &@@     \n"
+"                                                                               /@@        @&                     @@&    \n"
+"                                                                               ,@@@@@@@@@@@@@%                    @@    \n"
+"                                                                                  &@@@@     @@                    @@,   \n"
+"                                                                                   #@@  @@, @@*                   #@@   \n"
+"                                                                                       ,@@@@@@                    .@@   \n"
+"                                                                                  &@@&&&#*.                        @@   \n"
+"                                                                                  @@/                              @@   \n"
+"                                                                                   .@@@@@@.                        @@   \n"
+"                                                                                        @@                         @@.  \n"
+"                                                                                      /@@@                         @@*  \n"
+"                                                                                @@@@@@@*@@                         @@*  \n"
+"                                                                                       ,@@                         @@/  \n"
+"                                                                                   @@@@&@@.       @@@@@@@@         &@&  \n"
+"                                                                                @@@/    @@#      %@@    /@@        (@@  \n"
+"                                                                                         @@     ,@@      /@@       *@@  \n"
+"                                                                                        &@@@@@@@@@@/     .@@        @@  \n"
+"                                                                                        @@@#....,*@@      @@,       @@  \n"
+"                                                                                           .****..        @@@@@@@@@@@@  \n"
+"                                                                                           ,*(@@@@@@@@@@@@@@@@@@@@@&@@. \n"
+"@@@@@@@@@@@@@@@@@@@@*                              *#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout << "@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"       .*@@@@@@@@@@@@&(,                 /@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(          *%@@@@@@@@@@%  \n"
+"                                                                      .@@@@@@@@@@@@@@%                                  \n"
+"                                                                                .%@@@@@@@@@@@@@@@/                      \n"
+"                                                                           %@@@@@&.            .(@@@@@@@&               \n"
+"                                                                         @@@@   .@@@@@                  @@@(            \n"
+"                                                                        @@@     @@@@@@/                   @@@           \n"
+"                                                                       #@@                                 @@@          \n"
+"                                                                        @@@@@@@@@@@@@@&                    .@@.         \n"
+"                                                                           %@@@@    @@@                     @@@         \n"
+"                                                                                @@@@@@@                     @@@         \n"
+"                (#                                                         @@@@@@@@@@&                      @@@         \n"
+"               @@@@@@@@.                                                   @@@                             @@@          \n"
+" .....        ,@@@@@@@@@@@@@                                                 @@@@@@                        @@&          \n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@                                                    @@@                        &@@          \n"
+"              @@@@@@@@@@@                                               @@@@@@@,@@%                        %@@          \n"
+"              @@@@@@.                                                 &@@     #@@@                         @@@          \n"
+"                                                                               #@@                         @@%          \n"
+"                                                                            (@@@@@                        @@@           \n"
+"                                                                          @@@* %@@                        @@@           \n"
+"                                                                               &@@                       @@@            \n"
+"                                                                               @@@      @@@@@@@@@@,      @@@            \n"
+"                                                                               @@#     @@@      *@@@@@@@@@@@&           \n"
+"                                                                              @@@.    @@@       &@@@@@@@@@@@            \n"
+"                                                                            @@@@@@@@@@@@/                               \n"
+"                                                                            @@@@@@@@@@@@@  ,*(@@@@@@@@@@@@@@@@@@@@@#    \n"
+"@@@@@@@@@@@@@@@@@@@@*                              *#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"       .*@@@@@@@@@@@@&(,                 /@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(          *%@@@@@@@@@@%  \n"
+"                                                                      .@@@@@@@@@@@@@@@@@@@@@@@,                         \n"
+"                                                                    @@@@@/                   %@@@@*                     \n"
+"                                                                  &@@#    @@@@@&                 @@@                    \n"
+"                                                                 @@@      @@@@@@@                ,@@                    \n"
+"                                                                @@@        @@@&                  @@@                    \n"
+"                                                               *@@@@@@@@@@@##/                   @@@                    \n"
+"                                                                 ,@@@@@#((#&&@@@                 .@@/                   \n"
+"                                                                  ,@@@@      @@@                  @@@                   \n"
+"                                                                        (@@@ @@&                  .@@.                  \n"
+"                                                               @@@@@@@@@@@@@@@%                    @@,                  \n"
+"                                                  #@@@        %@@                                  @@*                  \n"
+"                              .@@&*               @@@@@@@      @@@#                                @@@                  \n"
+"                               .*#@@@@@@@@@@@@@@@@@@@@@@@@@@     ,@@@@@@@#                          @@&                 \n"
+"                                                 @@@@@@@@@@@@(          @@@.                        /@@.                \n"
+"                                                @@@@@@@@@@          /@@@@@@@@                        @@@                \n"
+"                                                                @@@@@/    &@@                        (@@                \n"
+"                                                               .&          @@&           @@@@@       (@@                \n"
+"                                                                       %@@@@@@         /@@# @@@      *@@                \n"
+"                                                                 &@@@@@@*  @@@        ,@@.  @@@      *@@                \n"
+"                                                                           &@@        @@@   @@@      *@@                \n"
+"                                                                           #@@@@@@@@@@@@    @@@      @@@                \n"
+"                                                                          .@@,      %@@@    @@@@@@@@@@@@                \n"
+"                                                                           .@@@@@@@@@,     .@@#      @@%                \n"
+"                                                                                           ,*#@@@@@@@@@@@@@@@@@@@@@#    \n"
+"@@@@@@@@@@@@@@@@@@@@*                              *#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"       .*@@@@@@@@@@@@&(,                 /@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(          *%@@@@@@@@@@%  \n"
+"                                                                      .@@@@@@@@@@@@@@%                                  \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                        #&@@@@@@@@.                     \n"
+"                                                                                 #@@@@@@@(.     .@@@@@                  \n"
+"                                                                             @@@@@,                 @@@                 \n"
+"                                                                         %@@@@                       @@@@@*             \n"
+"                                                                    /@@@@@*                          *@@ @@@@           \n"
+"                                           *@@@@@@@@@@@@@      (@@@@@(                       @@      @@@,  @@@          \n"
+"                                      &@@@@@#           @@@@@@@@*                     ,@@&          *@@@   .@@.         \n"
+"                                &@@@@@@.                   @@@            @@@@@@@@@@@@              @@@    #@@          \n"
+"                            @@@@@.  .       &@/            ,@@.                  *@@@@@             @@(    @@@          \n"
+"                         /@@@.        .@@@@                 @@#                 ...@###             @@@    @@@          \n"
+"                        @@@           ,@,  (@@             &@@                   /@                  @@@  @@@           \n"
+"                       %@@          @@                    @@@                              ,*(@@@@@@@@@@@@@@@@@@@@@#    \n"
+"@@@@@@@@@@@@@@@@@@@@*  @@@                         *#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(1000) );
+    cout << endl << endl;
+}
+
+
+void wumpus_eat_ani(){
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"    @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%,       .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"      %@@@@@@@@@@@@@@@@@@@@@@@@(                                 @@@@@@@@@@@@@@@@.                                      \n"
+"           #@@@@@@@@@@@@@@,                                      @@@@@@@@@@@@@.                                         \n"
+"                                                                  (@@@@@@&,       .......                               \n"
+"                                                                            @@@            ..*&@@@@@@%                  \n"
+"                                                                          @@                          #@@@              \n"
+"                                                                         @@     @@@@@@@                   @@            \n"
+"                                                                        &@*                                @@           \n"
+"                                                                        @@                                 @@           \n"
+"                                                                         @&                                *@(          \n"
+"                                                                         @@                                 @@          \n"
+"                                                                           .@@..&,.....@@//                 @@          \n"
+"                                                                            @@          @@                   @%         \n"
+"                                                                           @@.        %@                     @@         \n"
+"                                                                             @@@@@      @@                   @@         \n"
+"                                                                                @@      @&                   (@.        \n"
+"                                                                               @@@                           .@%        \n"
+"                                                                            @@@ @@                            @@        \n"
+"                                                                           @@   @@                            @@        \n"
+"                                                                         ,@&    @@                            @@        \n"
+"                                                                         @@     @@,                           @@        \n"
+"                                                                              @@@@/               *           @@        \n"
+"                                                                            &@% @@,          /@@/  @@         @@        \n"
+"                                                                           ,@(   @@         ,@#     @@   &@@&@@         \n"
+"                                                                                 @@@@&.     &@,     @@@@,    @%         \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@                         ,*,                          @@         *@@     .@&      @@(         \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@.              @@@@@@@@@@@@@@@@@@@@@@@@@%##/.   .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+   this_thread::sleep_for(chrono::milliseconds(1000) );
+    system("cls"); 
+    cout << "@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"    @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%,       .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"      %@@@@@@@@@@@@@@@@@@@@@@@@(                                 @@@@@@@@@@@@@@@@.                                      \n"
+"           #@@@@@@@@@@@@@@,                                      @@@@@@@@@@@@@.                                         \n"
+"                                                                  (@@@@@@&,       .......                               \n"
+"                                                                            @@@            ..*&@@@@@@%                  \n"
+"                                                                          @@                          #@@@              \n"
+"                                                                         @@      @@@@@@(                  @@            \n"
+"        @@@@@@@@@.                                                      &@*      @@@@@@&                   @@           \n"
+"     #@&         @@                                                     @@       %@@@@#                    @@           \n"
+"   .@@     @@@@& @@                                                      @&                                *@(          \n"
+"  %@    &@(     @@/                                                       *(@@@@&,                          @@          \n"
+", @@@&*@@         @&                                                               &@/.@@//                 @@          \n"
+" @@    @@         @#                                                           /@@@     @@                   @%         \n"
+"@      *@(       @@                                                        @@,        %@                     @@         \n"
+"         @@@(((@@@                                                           @@@@@      *                    @@         \n"
+"             @@   @@@@&                                                         @@                           (@.        \n"
+"           %@@@@,  (@  ,@@                                                    .@@@                           .@%        \n"
+"         @@. @%  @@@%    &@                                                .@@. @@                            @@        \n"
+"       @@/   @%   @@.@@@@ @%                                           @@@&     @@                            @@        \n"
+"    @@@      @*   @*     *@,                                       .#,        (@&@                            @@        \n"
+"           /@@   @@     @@@                                              .%@@@* #@,                           @@        \n"
+"         (@@.(@@ *@@@@@*                                              /*        &@/               *           @@        \n"
+"        @@,    #@(                                                              @@,          /@@/  @@         @@        \n"
+"   (@@@@         @&                                                              @@         ,@#     @@   &@@&@@         \n"
+"                 /@                                                              @@@@&.     &@,     @@@@,    @%         \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@                         ,*,                          @@         *@@     .@&      @@(         \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@.              @@@@@@@@@@@@@@@@@@@@@@@@@%##/.   .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout << "@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"    @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%,       .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"      %@@@@@@@@@@@@@@@@@@@@@@@@(                                 @@@@@@@@@@@@@@@@.                                      \n"
+"           #@@@@@@@@@@@@@@,                                      @@@@@@@@@@@@@.                                         \n"
+"                                                                  (@@@@@@&,                                             \n"
+"                                                                    *@@@@@@*       /@@@@                                \n"
+"                                                                 @@(   .@@*             @@                              \n"
+"                                                               @@     %@@@@@*            @@,                            \n"
+"            #@@@@@@@@@                                        @@      @@@@@@&              @@                           \n"
+"         @@/          @@                                     .@,      @@@@@@                @@                          \n"
+"       @@       .&@@@#.@@                                    %@                              @@                         \n"
+"      @@     (@@       @@                                    ,@&&&&&@@@@@@@@@@                @@                        \n"
+"     #@  @@@/@#        &@                                       @@@@@       .@,                @@                       \n"
+"   ( @@@@   @@        .@( %@@@,                                  ,@&    %@@@ @*                /@                       \n"
+"   /@@#     &@        @@   @@  @@.                                      @@@@@@                  @@                      \n"
+"    @(@/      @@@&(%@@(     @&   @@                               @@@@@@&,                       @#                     \n"
+"                 @@      @@@@@@,  @@                              @@                             .@%                    \n"
+"              ***(@@@@@@@%   ,@*  @@                                @@@,                           @@                   \n"
+"          ((**    @@          %@  @@                          @@.          @@                       @@                  \n"
+"                  @@           @@@@                               *&@@@@@@@@@#               ,      @@@                 \n"
+"                  @@         /                                              @@          @@%  ,@%     @@                 \n"
+"                  @@&                                             ...##&@@@@&@@        .@*    *@(     @@                \n"
+"                 @@/@.                                                       %@         @@     /@@@/***@@@              \n"
+"                #@  %@,                                                       @%         @&   @@     .&@@@              \n"
+"                @@    @@                                                      @@      ..  @%    .(((                    \n"
+"                @@     @@                                                      @@@@%     *@@%                           \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@                         ,*,                        .@(          &@                           \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@.              @@@@@@@@@@@@@@@@@@@@@@@@@%##/.   @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout << "@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"    @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%,       .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"      %@@@@@@@@@@@@@@@@@@@@@@@@(                                 @@@@@@@@@@@@@@@@.                                      \n"
+"           #@@@@@@@@@@@@@@,                                      @@@@@@@@@@@@@.                                         \n"
+"                                                                 *%@@@@@@&,                                             \n"
+"                                                       &@@@@&.(,  .,,&@@@@@%                                            \n"
+"                                @@@@@@*         #@@@#                      @@                                           \n"
+"                            .@@/       @@(    @@      @@@@@                 ,@%                                         \n"
+"         /@@@@@@@,         @@ (@@        ,@& ,@.     #@@@@@@@                 @@                                        \n"
+"       @@         @@/     @*     @@@       @*/@       @@@@@@@                  @@                                       \n"
+"     %@      @@@@@@@@               @@%    @@ @&                                @(                                      \n"
+"     @&    *@%     .@@                /@@ @@  /@.      .%@@@@@@                 @@                                      \n"
+"    #@    /@         @@                 @@(    ,@@@@@@,       @%                (@                                      \n"
+"    @@ /@@@@          @(    ,@                    ,@@@/       &@                .@.                                     \n"
+"    @@@    @@*       @@    @@                                  @/                @,                                     \n"
+"  /@@%        @@@@@@@@   @@                               &@@ @@                 @@                                     \n"
+"   @             @#    @@*                               ,@@@@@                   @@                                    \n"
+"     #@@@@@@@@@@@@@&@@@*                              @@,                          @@                                   \n"
+"                  /@                                 @@                            @@                                   \n"
+"                   @@                                *@@@@@@%                      %@                                   \n"
+"                    @@                      &@@@@@@@&&&&@@@@@                       @@                                  \n"
+"                     @@                                    @@       ,@@@@@@&        %@                                  \n"
+"                    @@@@#                     &%%%%%%@@@@@@@@      @@     .@#        @&                                 \n"
+"                   /@   @@                                 @@      @@      #@        @@                                 \n"
+"                   @@    .@@#                              #@&@@@@@@@*      @@        @@                                \n"
+"                   @&        @@                           %@        @@      &@@#     /@@                                \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@                         ,*,      @@@@@@@@@      &@          @&                               \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@.              @@@@@@@@@@@@@@@@@@@@@@@@@%##/.@@@,%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout << "@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"    @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%, #@,  .@&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"      %@@@@@@@@@@@@@@@@@@@@@@@@(         &@ .@&                  @@@@@@@@@@@@@@@@.                                      \n"
+"           #@@@@@@@@@@@@@@,             #@@@* .@@@@@@@@@@%.      @@@@@@@@@@@@@.                                         \n"
+"                                            @@              /@@@  (@@@@@@&,                                             \n"
+"                    @@@@*.      .@@@       @@    %@@@@@         (@@@                                                    \n"
+"                .@@.                @@      @@   @@@@@              @(                                                  \n"
+"             @@@@.@@@@@@@@. @@@@@@  @@       @@                    *@.                                                  \n"
+"         ,@@@  @%          @@     @@@@      /@@@@                  @@                                                   \n"
+"                          @@       @@       .@(  @@@,             %@                                                    \n"
+"                          @&       @@              @@             @@                                                    \n"
+"                          /@/     %@  *////  @@  @@.              @#                                                    \n"
+"                            %@@@@@@ @@      @@@@@@                @#                                                    \n"
+"                             @@    @@           .                &@                                                     \n"
+"                          .@@@@@@  @@.                          (@                                                      \n"
+"                        @@( %@  @@    /@@@                      @@                                                      \n"
+"                      /@&   @@%  &@      @*                     @@                                                      \n"
+"                    @@#     @&.@@@@@     @*                     @@                                                      \n"
+"                  @@       @@@     @@@@@@@                      @%                                                      \n"
+"                           @/.@@@       @#                     *@                                                       \n"
+"                          @@     .@@@@@@@/                     &@                                                       \n"
+"                          @@            @/            @@@@      @%                                                      \n"
+"                      @@@ %&@           @/         ,@@   @#     @&                                                      \n"
+"              /@@@@(         @@        *@         /@     @@@@@(  @@                                                     \n"
+"                              @@       *@         @@     @@     @@@                                                     \n"
+"                               @&      %@/@@@@@@@@@,       /@@@*                                                        \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@             @@        @@@*,                                                                  \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@.      @@     ,@@@@@@@@@@@@@@@@@@@@@@@@@%##/.   .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"    @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%,       .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"      %@@@@@@@@@@@@@@@@@@@@@@@@(                    ,@@@@@@*     @@@@@@@@@@@@@@@@.                                      \n"
+"           #@@@@@@@@@@@@@@,                       @@        *@@@@@@@@@@@@@@@@@.                                         \n"
+"                                                 @@     (@@@@.    (@@@@@@&,                                             \n"
+"                                                 @@     @@@@@@         @@&                                              \n"
+"                                  .$              @@     @@@@            @@                 @@&                         \n"
+"                                    ,%            @@@@@@@                 @(              *@@@@@                        \n"
+"                                       @@&                &@@            .@/  @@   (@@@@@.    @@                        \n"
+"                                         (@&     .@@@     .@*            @&    @@@            @@                        \n"
+"                               ..,%@@@@@@@@@&@@@@@  @@@@@(@@            @&      @@@(       %@@                          \n"
+"                                     @@%  ./* /   #@%    @@           @@             (@@@%                              \n"
+"                                    @@ &@%    (@@@#@    @@           @@                                                 \n"
+"                                   %@ #@          /@@ *@&           @@                                                  \n"
+"                               @@   @& .@@@          .@,           @@                                                   \n"
+"                               &@    @@    @@@.                   @@                                                    \n"
+"                                (@%    @@, @@.                   @@                                                     \n"
+"                                  /@@@    @@                    .@,                                                     \n"
+"                                      #@@@@                    /@(                                                      \n"
+"                                       ,@&                    &@                                                        \n"
+"                                      *@.          *         &@                                                         \n"
+"                                      @@        .@@@@       @@                                                          \n"
+"                                     @@/       @@  @@      &@                                                           \n"
+"                                    @@       @@    @@      @@                                                           \n"
+"                                 %@@(@@@@@ #@*     @@     @@                                                            \n"
+"                                @@        *@,     %@@@@@@@@*                                                            \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@      *@@@@&,   #@@   .@/,*,     @/                                                           \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@.              @@@@@@@@@@@@@@@@@@@@@@@@@%##/.   .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"/@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"    @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%,       .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"      %@@@@@@@@@@@@@@@@@@@@@@@@(                                 @@@@@@@@@@@@@@@@.                                      \n"
+"           #@@@@@@@@@@@@@@,                                      @@@@@@@@@@@@@.                                         \n"
+"                                          #@@@@@@@@&%*            (@@@@@@&,                                             \n"
+"                                       @@                     ,@@@                                                      \n"
+"                                     /@*       @@@@              @@.                                                    \n"
+"                                     @/       @@@@@@               @@                                                   \n"
+"                                    @@        @@@@@@                @@                                                  \n"
+"                       #@@*         @@         @@@@                 .@@                                                 \n"
+"                           @@@@.     #@@*                             @@                                                \n"
+"                                #@@@./(%&@@@@@@@@(//,                 /@                                                \n"
+"                              (@@@&   &@@@(                           .@,                                               \n"
+"                       #@@@%         @@                                @(                                               \n"
+"                                     ,@                               .@.                                               \n"
+"                                      @@,                             @@                                                \n"
+"                                        (@@(((@@,                     @#                                                \n"
+"                                        *%@@@@@@                      @#                                                \n"
+"                                   .@@@       @@                      @&                                                \n"
+"                                 @@          #@@                      @@                                                \n"
+"                               #@*       /@@% @%                      @@                                                \n"
+"                              .@       .@@   @@*                      @@                                                \n"
+"                                      #@.    @#          @@           @&                                                \n"
+"                                      @,    *@,         @@/@%         @&           @@                 @@                \n"
+"                                            %@        /@&   @@       *@.          @#@@@@@##&&.,,,##&@@@@@               \n"
+"                                            @@@@@@@% @@      @@      @@            /@@@,              @*                \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@                 #@      ,@@       @@     (*@@             ,@@             @@                 \n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@.           @@ @@@@@@@@@@@@@@@@@@@@@@@@@@##/.   .%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+
+    this_thread::sleep_for(chrono::milliseconds(1000) );
+    cout << endl << endl;
+}
+
+
+void pitt_fall_ani(){
+    system("cls");
+    cout << "                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"  .,*,                                                                                            /@@@@@@@@@@@@@@@@.    \n"
+"@@@@@@@@(                  /@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@        @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(      @@@@@@@@@@@(        %@@@@@@@@@@@@#       @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                              @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%                                                             ,@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                               @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*                                                               @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                *@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*                                                                @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@#                                                                   @@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                    #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                   #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                   @@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"      @@.          @@                                                                                                   \n"
+"        @@@*      %@.                                                                                                   \n"
+"            *@@@@@,    (*                                                                                               \n"
+"              *@         %@%                                                                                            \n"
+"              &@         (@@@                                                                                           \n"
+"            .@@@ ,@@@  @@@  @&                                                                                          \n"
+"           @@,*@     %@@    @&                                                                                          \n"
+"        &@@    @#  /@@@    .@(                                                                                          \n"
+"     @@@       @@@@%      /@,                                                                                           \n"
+"              @@#  #@@@@@@(                                                                                             \n"
+"             @@ .@@/                                                                                                    \n"
+"           &@/     @@(                                                                                                  \n"
+"          @@         @@,                                                                                                \n"
+"  .,*,   @@            @@                                                                         /@@@@@@@@@@@@@@@@.    \n"
+"@@@@@@@@@.              (@(/@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@        @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(      @@@@@@@@@@@(        %@@@@@@@@@@@@#       @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                              @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%                                                             ,@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                               @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*                                                               @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                *@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*                                                                @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@#                                                                   @@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                    #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                   #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                   @@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"                                    /@            @@                                                                    \n"
+"                                     @@.         @@                                                                     \n"
+"                                       ,@@@@@@@@@@     @@@%                                                             \n"
+"                                            @(            @@@@.                                                         \n"
+"                                           /@@@@,        .@*  @@                                                        \n"
+"                                      /@@@/@@    *@@@@(  @@    @@                                                       \n"
+"                                #%%*,      @&          .@@     &@                                                       \n"
+"                                          @@           %@      /@                                                       \n"
+"                                        @@@@*         @@&      @@                                                       \n"
+"                                      &@    @@       @@     *@@*                                                        \n"
+"                                     @@      @@                                                                         \n"
+"                                   @@         @@                                                                        \n"
+"                                 (@@           &@                                                                       \n"
+"  .,*,                        .@@@              @@                                                /@@@@@@@@@@@@@@@@.    \n"
+"@@@@@@@@(                  /@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*  @@@@@@@&@@@@@@@  .@@@@@@        @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@  .@@@@*       @@@/     @@(        %@@@@@@@@@@@@#       @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                              @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%                                                             ,@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@              @@@@                                             @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*               &(                                              @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                              #(                                *@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                             (@@#                                @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*             @&                                                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@            &@@@(                                               @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@#                                                                   @@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                    #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                   #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                   @@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                           *@   ,@@*       %@@@                                                         \n"
+"                                         *@@@%(&@@@.     (@@@*#(@,         @@@@@/                                       \n"
+"                                                      @@#  @@,(@@@@    %@@       @@                                     \n"
+"                                                         *@#       @@@@@#.        @@                                    \n"
+"                                                         @@         @@     /@@@@@@#@@                                   \n"
+"                                                        (@          #@              @/                                  \n"
+"  .,*,                                                   &@%        @@  @@                        /@@@@@@@@@@@@@@@@.    \n"
+"@@@@@@@@(                  /@@@@@@@@@@@@@@@@@@     @@@.     .%@@@@@&  ,@@        %@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&   .@@@*   &@     @@         ,@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@,  @@@@@@    @@@@@.      @@@.@@  @@          &@@       .@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                /@@@@           @@                            @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%                                 @@                          ,@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@              *&                    @@###.                     @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*              #@@@                 *@     &@@*                 @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                   /@%         .@@              *@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                  @@                             @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                &@,                     ,@@#     @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*                            *@@@                        @@@     @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@         @@*                             */              @      @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@          @@@#                      *@@@@@@@@@@.                 @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@#           @%                            *#.                       @@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                                    #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                       @@@@@@@@@@#                                 #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                            %@@@@@@                                @@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"                                                                                                                        \n"
+"  .,*,                                                                                            /@@@@@@@@@@@@@@@@.    \n"
+"@@@@@@@@(                  /@@@@@@@@@@@@@@@@@@                                   %@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&                             ,@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@,  @@@@@@    @@@@@.                          &@@       .@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                /@@@@                                         @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%                                                             ,@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                                                               @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*                *@                                             @@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                @@@@                                            *@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                         &@@@@@@@@@@&                            @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                      @@           @@@@                          @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*                          &@@%@@*                @@(   @@@*     @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                            *@                   @@&&.     @@&  @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                   *,              @@@    @@@        (@@#    @@  @@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@#                    @@@@           @@        @@           @@@ &@   @@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@                            @@,    &@          @@   @@       .@@@   #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                             @@    @@*        @@  @@          @@   #@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@                               @@     %@@@@@@@&  @@                @@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((\n"
+"((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((*#######,/((((((((((((((((((((((((((((((((((((((((\n"
+"(((((((((((((((((((/#################################,(((((((((,##########################,****/((((((((((((((((((((((((\n"
+"(((((((((((((((((#############(//////########################################//(########*###########((((((((((((((((((((\n"
+"(((((((((((((((((########/####,       /((##############################(((,              ,###(########*(((((((((((((((((\n"
+"(((((((((((((((((####(##.                                                                     ##########((((((((((((((((\n"
+"(((((((((((((((/######/                              .//.                                        (#(####,(((((((((((((((\n"
+"((((((((((((((*#######                            @@@@#   %@@@@*                                   ##*###/((((((((((((((\n"
+"((((((((((((((####(##                            @@*            /@@@                                 #####((((((((((((((\n"
+"((((((((((((*#####/#                           &@(                  @@                                #####(((((((((((((\n"
+"((((((((((((######/#                          @@                     %@(                               #####((((((((((((\n"
+"(((((((((((((#####/#                        .@(                       *@,                              #####,(((((((((((\n"
+"(((((((((((((#####/#                        @%                         @@                              (#####/((((((((((\n"
+"(((((((((((((#####/#                       %@                          @@                               #(####*(((((((((\n"
+"(((((((((((((#####(#                       @@                         %@                                .#/####*((((((((\n"
+"(((((((((((((*####(#                       @@                        .@,                                 *#/####((((((((\n"
+"((((((((((((,#######                        @@                       @@                                   ######((((((((\n"
+"(((((((((((/#######                           @@*                  @@@                                   ######/((((((((\n"
+"((((((((((/#######                               ,@@@@@@@@@@@@@@@@%                                     #######(((((((((\n"
+"((((((((((/####(#/                                       @@                                            #######*(((((((((\n"
+"(((((((((((###/##                           .&@@@@@@@@@@@@@@@@@@@@@@@@                                 #,####,((((((((((\n"
+"(((((((((((##*#.                                       %@@*                                           ######((((((((((((\n"
+"((((((((((/##/#                                     .@@   %@&                                        /#*##*(((((((((((((\n"
+"((((((((((*#####                                            *                                       ######/(((((((((((((\n"
+"(((((((((((######                                                                                (#########(((((((((((((\n"
+"((((((((((((((###*###(                                                    ##############(///(##############(((((((((((((\n"
+"((((((((((((((((#########/////*######(                        *(#######/*###############################/(((((((((((((((\n"
+"(((((((((((((((((#####################(*/////////////*/(################################,(((((((((((((((((((((((((((((((\n"
+"((((((((((((((((((((((((((((((((((((((/(###################/((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((\n"
+"((((((((((((((((((((((((((((((((((((((((((((((/(##(,((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((\n"
+"((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((((\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&&&&@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&%&&&&&&&&&%((((((%%%%%%(%%%%&&&&&&&&&&&&&(&&&&&&&&&&&&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@&&&&&&&#&/%%%%%%*%%%%%%%%%%%%%/..../%%%%%(#%%%%%%%%%%%&&&&&&&&&&&&&&&%%&&&&&&&&&&@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&%%%%(%%#               @@@@@&       /(%%%%%%%%#%%%%%%%%%%%%%%%*&&&&&&&&&&@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@&&&&&&&&&%%*%%                @@#      @@                 *%%%%%%%%%%%%(%%&&&&&&&&&&@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@&&&&&%&#%%#%                *@*           &@@&                       %%%%%%&&&&&&&&&@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@&&&&&&&%%*%%                &@                /@,   .@@(              %/%%%&&&&&&&&&@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@&&&&(&&%#%                 .@*                 @#   %@  #@            %(%%&&&&&&&&&@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@&&&&&&&&%#%#                /@*                ,@.   @&   *@           %(%%&#&&&&&&&@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@&&&&&&&(&#%%%                  @@#             *@/    @*    &@          %#%%&&&&&&&&&&@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@&&&&&&&(&#%%%#                    &@@@@%///%@@@         %@@( @          %%%%/&%&&&&&&&&@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@&&&&&&&/&#%%*%,                 %@@@/   @*  &@@@*          ,@@          #%%%%&&&&&&&&&&&@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@&&&&&&&&&&%%%#%                       (@@,                @@@(          %#%%%#&&&&&&&&&&@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@&&&&&&&&&#%%%/%%                      @@                  @&          %%%%%%%&&&&&&&&&@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@&&&&&&&&&&&&%%%%%%                    @@ @@&                          ,%(%%%&&&&&&&&&&&@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@&&&&&&&&&&&&%&&%%%#%%#               @@     %@#                   *%%%%/%%%%&#&&&&&&&&&@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@&&&&&&&&&&&&&&&&&&&%%%%(%/                                     .%%%%%%/&&&&&%&&&&&&&&&@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@&&&&&&&&&&&&&&&&&&&%%%%%%%#######/.                 (%%%%%%#//%%%/&&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@&&&&&&&&&&&&&&&&%%%%%%%%%%%%%%%%%%%%%%%%((#%%%%%%%%%%%(#%&&&&%&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&&&&&&&&&&&%#%%%%%%%%%&&&&&&&&&&&&&&&&&(&&&&&&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&@@@@@@@@@@@@@@@@@@@@&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&/@@@@@@@@@@(#%%%#@@@&@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@&@/&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&/@@(@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@(@#&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&*&&&&&&&&&&&&&&&&&&&&&&&&@#@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&&&&%&&(%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%(&&&*&&&&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@*@(&&&&&&&&&&&&&%%%%%%%%%%%%%%%%%%######%%%%##%%%%%%%%*%%%%%%%&&&&&&&&&&&%@&@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&%&&/%%%%%%%(%%%%###/, .//.           %%%%(((((%%%%%%%%%#&(&&&&&@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&(&(%%%%%%%%%%%          #(   ##                  ,%%#%%%%%%&%&&&&#@(@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@#@&&&&&&&&&%%%%%%(%,           /#    *#                   .%/%%%%%(&(&&&&&@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@&@&&&&&&&&&&%%%%%(%             *#####*                    %%%%%%%%*&(&&&&&@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@%@#&&&&&&&&%%%%%%(%               /###/.  ## %/           *%#%%%%%%&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@%@/&&&&&#&(%%%%%%#%%#             ####    *%*             #%%%%%%%&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@#@%&&&&&/&%%%%%%%%%%            ##  .#,                   %#%%%#&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@#@#&&&&&&&&&#%%%%%*%*                                    %%%%%,&(&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&&#%%%%%%%(%%%,                          #%%%( #&&&&&%&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&&&%%%%%%%%%%%(%%%                    /%%/%%%&%&&&&&&&&&&&&&@@#@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&&&&#&&/%%%%%%%%%%%%%%//////%%%%%%%%%%%%%%#&&&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@#@/&&&&&&&&&&&&(&&&&&&&&&&&&&&&&&&/%%%%%%%%%(&&&##&&&&&&&&(@@@@@@@(#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(&&&&&&&&&&&&&&&&&&&&&&&&&#((#&&&&#(&&&&&&&&&&#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(@@(&@@@@@@@@@@@@@@@@@@@@@@@#&&&&&&&&&&&&&&%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%@&@@@@@@@@@@@@@@@@@@@@@@@@&(@@@@@@@@@@@@(@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&@@@@@@@@@@@@@@@@@@@@&#(@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@#@@%&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&*@@@@@@@****@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@#@%&&&&&&&&&&&&&&&&&&&&#&&&&&&&&&/&&&&&&&&&&&&&&&(@@&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@#@#&&&&&&%&&&&&&&%#((/#%%((((((%%%%%%##&&&&(///#&&&&&&&%@%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@#@&&&&&&&&&%%%%%%%%%%%/%(         *%%%#%%%%%%%%%%&#&&&&&&&#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&#&/%%%%%%%%%%%%    #( )#       %%%%((#%%%%(&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@%&&&&&&&#%%%%%%%%%%%.    .*##*.            %/%%%%%&*&&&&@@%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&/&%%%%%%%%%%#      .##.           %%%%%%%#%&#&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&@&&&&&&&(%%%%%%%%%/     .   .      .%%%%(%%%%%&%&&&&&&@&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&&&&&&%%%%%%%%%%%%# ((        #%%%%%%%%%%%(&#&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@#@&&&&&&&/%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#&&(&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@#&&&&&&&&&&&&*%%%%%%%%%%%%%%%%%%%%%%%%&&%&&&&&&@@@@&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@#&&&&&&&&&&&&&&&&&&&&&&&%%(&&&&&%&&&#&&&&&(@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(%&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@(@#&@@@@@&%%%%%%%%#@@@@@#%&&&&%%@@%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    cout << endl << endl;
+}
+
+void empty_arrows_ani(){
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 &@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&.              %@@@@@@@@@@@@@@@@@@@@@@@@@@&                            ,\n"
+"                                                                         ##%%#*                                         \n"
+"                                                                                                                        \n"
+"                          #@&.      /@@                                                                                 \n"
+"                      #@                @@                                                                              \n"
+"                    @@        (%%%%%(     /@                                                                            \n"
+"              @ #@@    ,@@@(             @# @                                                                           \n"
+"             @@ @        @                 @@                                                                           \n"
+"                        @                   @                                                                           \n"
+"                       %                    @                                                                           \n"
+"                       #@                 #@                                                                            \n"
+"                          @             @@                                                                              \n"
+"                            (@@@((#@@*                                                                                  \n"
+"                                &@#     @@                                                                              \n"
+"                             @   @ @@   &%  @%                                                                          \n"
+"                             @   @     @@    &(                                                                         \n"
+"                              @@ @   &@@ @@@  @   /@                                                                    \n"
+"                                 @    @      ,@  @  .@                                                                  \n"
+"                                 @   @      #@                                                                          \n"
+"                                 @  (@@@@@@*                                                                            \n"
+"                               ,@  &@                                                                                   \n"
+"                              &#     ,@                                                                                 \n"
+"                             @        .@                                                                                \n"
+"                            @          @                                                                                \n"
+"                          @(           @                                                                                \n"
+"                         @            @        /%##/*,,,, .,*&@@@@@@@@@@@@@@&                                           \n"
+"       ,,&@@@&      #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(1000) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 &@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&.              %@@@@@@@@@@@@@@@@@@@@@@@@@@&                            ,\n"
+"                                                                         ##%%#*                                         \n"
+"                           .(@@@&(                                                                                      \n"
+"                     ,@@@@*        &@@@@                                                                                \n"
+"                  (@@.                  /@@,                                                                            \n"
+"              @@ @@       (@@#      .%@@@ .@@                                                                           \n"
+"            @@@@@# *@@@& @@              @@@@                                                                           \n"
+"          /@@@@         @@                @@@                                                                           \n"
+"                        @@                @@                                                                            \n"
+"                        @@*              @@                                                                             \n"
+"                         .@@           @@@                                                                              \n"
+"                            &@@@# (@@@@@@@@@,                                                                           \n"
+"                                 %@, @@     @@.                                                                         \n"
+"                        &@@@(.  .,@@ @@      /@&                                                                        \n"
+"                       @@        .@@@@@       @@                                                                        \n"
+"                       @@@@@@@&&&&@@@@@     ,@@@  @@@                                                                   \n"
+"                                 ,@( @@       @@ &@@&                                                                   \n"
+"                                  @@ &@       @@                                                                        \n"
+"                                  @@ /@*     @@*                                                                        \n"
+"                                  @@&@@@@@@@@                                                                           \n"
+"                                %@&&@@                                                                                  \n"
+"                               %@*   @@                                                                                 \n"
+"                              @@,     .@@                                                                               \n"
+"                             @@         @@                                                                              \n"
+"                            @@.         /@,                                                                             \n"
+"                           *@%          &@.    /%##/*,,,, .,*&@@@@@@@@@@@@@@&                                           \n"
+"       ,,&@@@&      #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(1000) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 &@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&.              %@@@@@@@@@@@@@@@@@@@@@@@@@@&                            ,\n"
+"                                                                         ##%%#*                                         \n"
+"                                                                                                                        \n"
+"                        #@@@@@@@@@@@&                                                                                   \n"
+"                    %@@#             ,@@@                                                                               \n"
+"                  @@&     &@@@*     /@@@&@@                                                                             \n"
+"                @@%     *@%             @@@@                                                                            \n"
+"               @@   %@@@@@               @@                                                                             \n"
+"              &@@@@/   .@%               @@                                                                             \n"
+"           @@@@@        @@               &@(                                                                            \n"
+"           @@@@@@       .@@             @@@,                                                                            \n"
+"           .@@#            @@@@&&&&@@@@@@@  @@@                                                                         \n"
+"           .*                    @@   (@#     @@/                      @@@@@%                                           \n"
+"                            &@@@@@@@   @@      #@(            .,,*////(@@@@@@@@@@&                                      \n"
+"                       ,@@@#     .@&@@ #@*      @@     ,,.             @@@@@@@@@                                        \n"
+"                     %@@         (@,  #@@@      @@                     @@@@&                                            \n"
+"                    @@           @@      @@    #@/                                                                      \n"
+"                                 @@      @@ *@@@                                                                        \n"
+"                                 @@   .*@@@@                                                                            \n"
+"                                 ,@%                                                                                    \n"
+"                                 @@@@                                                                                   \n"
+"                                @@  @@*                                                                                 \n"
+"                               @@    .@@                                                                                \n"
+"                              @@       @@                                                                               \n"
+"                             @@         @@(                                                                             \n"
+"                            @@           *@@   /%##/*,,,, .,*&@@@@@@@@@@@@@@&                                           \n"
+"       ,,&@@@&      #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 &@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&.              %@@@@@@@@@@@@@@@@@@@@@@@@@@&                            ,\n"
+"                                                                         ##%%#*                                         \n"
+"                                                                                                                        \n"
+"                       &@@@@@@@@@@@@@*                                                                                  \n"
+"                   &@@#      (&@@@#   @@@,                                                                              \n"
+"                 #@@      @@(       %@@@,@@                                                                             \n"
+"                @@      *@&             @@@&                                                                            \n"
+"               @@     /@@@               @@                                                                             \n"
+"              ,@%#@@@@  @@               *@/                                                                            \n"
+"           @@@@@@       %@/              @@                                                                             \n"
+"           .@@@@@@       ,@@            @@@@@%                                                                          \n"
+"           %@&@@            @@@@@@@@@@@@@     @@                                                         @@@            \n"
+"           @*                    (@,    @@     @@.                                                       @@@@@@@@.      \n"
+"                   &.            (@,    %@@     @@                                    @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@   \n"
+"                    *@@@@&&&&@@@@@@@@@@&  @@   &@.                                                       @@@@@@@@.      \n"
+"                                 @@        @@ @@.                                                        @@@@           \n"
+"                                 @@         @@@                                                                         \n"
+"                                 @@   .@@@@@/                                                                           \n"
+"                                 (@&                                                                                    \n"
+"                                 @@@@                                                                                   \n"
+"                               *@&  #@%                                                                                 \n"
+"                               @@     @@                                                                                \n"
+"                              @@       @@                                                                               \n"
+"                             /@(        @@                                                                              \n"
+"                             @@          @@                                                                             \n"
+"                            @@           @@    /%##/*,,,, .,*&@@@@@@@@@@@@@@&                                           \n"
+"       ,,&@@@&      #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 &@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&.              %@@@@@@@@@@@@@@@@@@@@@@@@@@&                            ,\n"
+"                                                                         ##%%#*                                         \n"
+"                            ,(##(*.                                                                                     \n"
+"                      ,@@@@*        #@@@                                                                                \n"
+"                   @@@      .@@@@@@@@(  @@                                                                              \n"
+"                 &@%      @@*         @@@@@                                                                             \n"
+"               &@@       @@              @@&                                                                            \n"
+"              @@        /@/              @@                                                                             \n"
+"             @@&&&(.    @@               @@                                   #@@@@@&                                   \n"
+"          @@@@@         @@               @@                                    @@@@@                                    \n"
+"        @@&@@  *&        @@             @@*.                                    @@@                                     \n"
+"         ,@@@@            ,@@@&/     *@@& @@(@@@.                                                                       \n"
+"                                 /@@(    ,@#   /@#                                                                      \n"
+"                            #&&&/&@.     @@     @@,                             &@                                      \n"
+"                       @@@%      @@@@@@@@@&      @@                            .@@                                      \n"
+"                     ,@@         @@     @@       @@                                                                     \n"
+"                     ,(         @@      @@       @@                                                                     \n"
+"                                .@#     @@     &@@                                                                      \n"
+"                                *@/   (@@@@@@@&                                                                         \n"
+"                               @@@@                                                                                     \n"
+"                             .@@  @@                                                                                    \n"
+"                            *@%   &@                                                                                    \n"
+"                           .@@    .@%                                                                                   \n"
+"                           @@      @@                                                                                   \n"
+"                          @@       %@.                                                                                  \n"
+"                         (@/        @@         /%##/*,,,, .,*&@@@@@@@@@@@@@@&                                           \n"
+"       ,,&@@@&      #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 &@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&.              %@@@@@@@@@@@@@@@@@@@@@@@@@@&                            ,\n"
+"                                                                         ##%%#*                                         \n"
+"                          .@@@@@@@@@&*                                                                                  \n"
+"                       @@@.           *@@@@                                                                             \n"
+"                     @@*    @@@@@@@@,      @@,                                                                          \n"
+"                   &@&    @@.        (@@@   &@,                          .%@@@@@&%%/.                                   \n"
+"                  @@     @@             @@  #@,                         *@@@@@@@@@@@@@@@                                \n"
+"                @@&  ,(#(@,              @@ @@                           @@@@@@@@@@@@@@#                                \n"
+"           .@@&@@@@/    %@.              .@@@                             %@@@@@@@@@@@&                                 \n"
+"         @@ *@@*         @@               @@                               ,@@@@@@@@@&                                  \n"
+"          @@@             @@             @@                                 *@@@@@@@                                    \n"
+"          . @.             @@@@         @@.,                                 .@@@@                                      \n"
+"                               *@@@@@@@@*  @@.&@@                                                                       \n"
+"                                  @@      @@    @@                                                                      \n"
+"                            (@@@@@@@@@@  @@      @%                          ,%@.                                       \n"
+"                      .@@@&     @@     @@@.      @%                         &@@@@                                       \n"
+"                               @@       %@@&    @@                           .,                                         \n"
+"                              @@       &@(     @@                                                                       \n"
+"                             ,@@     @@@@@@@@@@.                                                                        \n"
+"                            @@@                                                                                         \n"
+"                           @@ @@                                                                                        \n"
+"                          &@   @@                                                                                       \n"
+"                         ,@&    @@.                                                                                     \n"
+"                         @@      @@                                                                                     \n"
+"                        @@.       @@                                                                                    \n"
+"                        @@        @@           /%##/*,,,, .,*&@@@@@@@@@@@@@@&                                           \n"
+"       ,,&@@@&      #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(700) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 &@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&.              %@@@@@@@@@@@@@@@@@@@@@@@@@@&                            ,\n"
+"                                                                         ##%%#*                                         \n"
+"                                                                                                                        \n"
+"                                                                     .@%                                                \n"
+"                                 &@@@@@@@@@@@@@(                   @@%                                                  \n"
+"                             /@@@               @@@              @@@@        (%%%%%#*                                   \n"
+"                           @@&                    @@            @@          #%%%%%%%%%%                                 \n"
+"                         @@(         #@@@*.,#@@@@. @@         .@@           %%%%%%%%%%                                  \n"
+"                   /@% %@@@@@@@@@@@@@@           @@(@#                     ,%%%%%%%%%                                   \n"
+"                 /@@ @@@           %@,            @@@                      #%%%%%%%%                                    \n"
+"                   (@@ @@          @@             @@                       %%%%%%%,                                     \n"
+"                   @.              &@             @@                        %%%%%                                       \n"
+"                                    @@           @@                                                                     \n"
+"                                     %@&      /@@*                         %%%%%%                                       \n"
+"                                     .@@@@@@@@,                            %%%%%%.                                      \n"
+"                               @@@@@@@@@@                                               @@*&@@                          \n"
+"                               @@ &@/    /@@@                                            &@@                            \n"
+"                                @@@         @@#                                       @@@(                              \n"
+"                                @@%@@.       .@@                                                                        \n"
+"                               @@    @@                .@(                                                              \n"
+"                              @@                      ,@@                                                               \n"
+"                             @@@(       .@.    #@@@@&* @@                                                               \n"
+"                            @@ (@%       .@@@%.      (@@                                                                \n"
+"                          /@@   @@*         .@@@%%@@@%                                                                  \n"
+"                         @@.     @@                                                                                     \n"
+"                        @@        @@           /%##/*,,,, .,*&@@@@@@@@@@@@@@&                                           \n"
+"       ,,&@@@&      #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 &@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&.              %@@@@@@@@@@@@@@@@@@@@@@@@@@&                            ,\n"
+"                                                                         ##%%#*                                         \n"
+"                                                                           ,####(                                       \n"
+"                                                                      &@@@,               (                             \n"
+"                                                                   @@@                ,&&&&&                            \n"
+"                                                                 @@@@@             %&&&&&&&&&                           \n"
+"                                                                @@   .@@           .&&&&,(&&&                           \n"
+"                                                               @@      #@@        &&&&&&&                               \n"
+"                                                              @@         *@@    %&&&&&&&/                               \n"
+"                                                @@@@@        .@&               %&&&&&&%         .@%                     \n"
+"                                                &@ @@&@@@@@  %@.              %&&&&&&#           @&                     \n"
+"                                                       %@( &@@*              &&&&&&&/            @%                     \n"
+"                                                        @@    @@@          ,#&&%/      %#       @@                      \n"
+"                                                      #@@@@@.   @@        &&&&&&&&     /@&     %@,                      \n"
+"                                                   (@@,     %@@  @@      *&&&&&&&        @@   &@(                       \n"
+"                                                  @@         ,@&  @@                      @@ @@                         \n"
+"                                                 .@%          @@  @@                     .@@@                           \n"
+"                                                  @@          %@ @@                   @@@(                              \n"
+"                                          @@@@@@.@@@@         @@@@              @@@@@/                                  \n"
+"                                         /@( *@@@@*  /@@@@@@@@*                                                         \n"
+"                                          &@@@   .@&                                                                    \n"
+"                                       @@@ ,@@    #@,                                                                   \n"
+"                                /@@@@@#     #@#    @@                                                                   \n"
+"                            .@@@ ,@@         @@    @@                                                                   \n"
+"                         %@@#   @@,                (@                                                                   \n"
+"                       @@,    (@@              /%##/*,,,, .,*&@@@@@@@@@@@@@@&                                           \n"
+"       ,,&@@@&      #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(500) );
+    system("cls");
+    cout <<"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+" @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"                 &@@@@@@@@@@@@@@@@@@@@@@@@@@@@@&.              %@@@@@@@@@@@@@@@@@@@@@@@@@@&                            ,\n"
+"                                                                         ##%%#*                                         \n"
+"                                                                           *&@@@@@@@@@@@,                               \n"
+"                                                                     @@@@@,            &&@@@&                           \n"
+"                                                                  @@@               (&&&&   &@@                         \n"
+"                                                                 @@@@@        *&.&&&&&&&&&    @@                        \n"
+"                                                                 @@  .@@        &%&&&&&&&&&    @@                       \n"
+"                                                                @@      @@.     &&&&&&&&&&&*    @@                      \n"
+"                                                               @@.        @@  (&&&&&&&&         @@                      \n"
+"                                                               @@           @@@&&&&&&&          %@.                     \n"
+"                                                               @@          #&&@@@&&&.           ,@(                     \n"
+"                                                               @@         &&&&&&&@@,            @@                      \n"
+"                                                               @@        &&&&&&&&&/@@.          @@                      \n"
+"                                                .@@  @          @@     ,&&&&&&&&&    @@@       @@                       \n"
+"                                               #@@@@@@@@@@@#     @@   &&&&&&&&&.       *@@    @@                        \n"
+"                                                    (  @@/   @@@@ @@, &&&&&&&&           /@@@@%                         \n"
+"                                                        (@(      &@@@@*  %&&            .@@@@                           \n"
+"                                                         @@@@@@    @@(,@@@%         #@@@@                               \n"
+"                                                       @@&     .@@@ /@(    &@@@@@@@&                                    \n"
+"                                                      #@.         @@ @@                                                 \n"
+"                                                 (%/  #@.         #@*@%                                                 \n"
+"                                                @@ /@@@@@&        @@@@                                                  \n"
+"                                         *@@@@@@@@   @@  &@@@@@@@@/                                                     \n"
+"                                   &@@@@@,       @@   @@                                                                \n"
+"                               #@@@  @@@         @@   #@.                                                               \n"
+"                           .@@@     @@         /%#@@*,&@, .,*&@@@@@@@@@@@@@@&                                           \n"
+"       ,,&@@@&      #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n"
+"@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n";
+    this_thread::sleep_for(chrono::milliseconds(1000) );
+    cout << endl << endl;
+
+
+
+
+
+
+}
