+++ animations.hpp
@@ -1,0 +1,4 @@
+void kill_wumpus_ani();
+void wumpus_eat_ani();
+void pitt_fall_ani();
+void empty_arrows_ani();
